+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/sections/AbstractKlassApiSections.java
@@ -1,0 +1,6 @@
+package no.ssb.klass.api.migration.dataintegrity.sections;
+
+import no.ssb.klass.api.migration.dataintegrity.AbstractKlassApiDataIntegrityTest;
+
+public class AbstractKlassApiSections extends AbstractKlassApiDataIntegrityTest {
+}
