+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/versions/KlassApiVersionByIdTest.java
@@ -1,6 +1,7 @@
-package no.ssb.klass.api.migration.dataintegrity;
+package no.ssb.klass.api.migration.dataintegrity.versions;
 
 import io.restassured.response.Response;
+import no.ssb.klass.api.migration.dataintegrity.AbstractKlassApiDataIntegrityTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
@@ -11,14 +12,7 @@ import static no.ssb.klass.api.migration.MigrationTestConstants.*;
 import static no.ssb.klass.api.migration.MigrationTestUtils.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class KlassApiVersionByIdTest extends AbstractKlassApiDataIntegrityTest{
-
-    static Integer randomId;
-
-    @BeforeAll
-    static void beforeAllVersions() {
-        randomId = generateRandomId(2000);
-    }
+public class KlassApiVersionByIdTest extends AbstractKlassApiVersions {
 
     @Test
     void getVersionById() {
